docker/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
torch==2.1.0
torchvision==0.16.0
python-multipart==0.0.6
python-doctr[torch]==0.7.0
easyocr==1.7.0
paddlepaddle-gpu==2.5.1
paddleocr==2.7.0.3
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
rapidfuzz==3.5.2
python-Levenshtein==0.23.0
pydantic==2.5.0